pyproject.toml
[tool.poetry]
name = "photocrowd_tech_test"
version = "0.1.0"
description = ""
authors = ["Tom Morledge <morledge.t@outlook.com>"]

[tool.black]
skip-string-normalization = true
target-version = ["py38"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
  | venv
)/
'''

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db"
python_files = ["tests.py", "test_*.py"]
testpaths = [
    "leaderboard"
]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.1.13"
django-environ = "^0.7.0"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.10.0"
whitenoise = "^5.3.0"
argon2-cffi = "^21.1.0"
pytz = "^2021.3"
django-filter = "^21.1"
django-utils-six = "^2.0"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]
Werkzeug = "^2.0.2"
mypy = "^0.910"
django-stubs = "^1.9.0"
pytest = "^6.2.5"
pytest-sugar = "^0.9.4"
flake8 = "3.9.2"
flake8-isort = "^4.0.0"
coverage = "^6.0.1"
black = "^21.9b0"
pylint-django = "^2.4.4"
pre-commit = "^2.15.0"
factory-boy = "^3.2.0"
django-debug-toolbar = "^3.2.2"
django-extensions = "^3.1.3"
pytest-django = "^4.4.0"
django-coverage-plugin = "^2.0.1"
Sphinx = "^4.2.0"
sphinxcontrib-django = "^0.5.1"
sphinx-autobuild = "^2021.3.14"
bandit = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
